[Refactor]: 스크랩 삭제 여러 개 가능하도록 수정
