+++ src/main/java/com/poppy/domain/scrap/dto/request/UserScrapReqDto.java
@@ -1,0 +1,10 @@
+package com.poppy.domain.scrap.dto.request;
+
+import lombok.Getter;
+
+import java.util.List;
+
+@Getter
+public class UserScrapReqDto {
+    private List<Long> scrapIds;
+}
